requirements.txt
Django==1.10
Jinja2==2.9.5
MarkupSafe==1.0
Pillow==4.1.1
chardet==2.2.1
colorama==0.2.5
command-not-found==0.3
configparser==3.5.0
coreapi==2.3.0
coreschema==0.0.4
django-ckeditor==5.3.0
django-js-asset==0.1.1
django-rest-swagger==2.1.2
djangorestframework==3.6.2
html5lib==0.999
itypes==1.1.0
language-selector==0.1
mysqlclient==1.3.10
olefile==0.44
openapi-codec==1.3.1
pycurl==7.19.3
pygobject==3.12.0
python-apt==0.9.3.5ubuntu2
requests==2.13.0
simplejson==3.10.0
six==1.5.2
ufw==0.34-rc-0ubuntu2
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.7.1
wheel==0.24.0
